code refactor and bug fix
